UPDATE: generic form modal added, update all components page
