+++ src/common/components/modal/use-modal.hook.js
@@ -1,0 +1,41 @@
+'use client';
+
+import { yupResolver } from '@hookform/resolvers/yup';
+import { useEffect, useState } from 'react';
+import { useForm } from 'react-hook-form';
+
+export default function useModal(validationSchema, show) {
+  const [open, setOpen] = useState(false);
+
+  useEffect(() => {
+    setOpen(show);
+  }, [show]);
+
+  const {
+    register,
+    handleSubmit,
+    setValue,
+    formState: { errors }
+  } = useForm({
+    resolver: yupResolver(validationSchema)
+  });
+
+  const handleClickOpen = () => {
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
+  return {
+    open,
+    setOpen,
+    register,
+    handleSubmit,
+    setValue,
+    errors,
+    handleClickOpen,
+    handleClose
+  };
+}
